Procedure updateStocksTrigger working, successfully added to db, made minor corrections to java code

+++ src/fincrawl/BuySellTriggers.java
@@ -16,11 +16,11 @@ public class BuySellTriggers {
         System.out.println("Price and Trigger Values for " + symbol);
         System.out.println("*******************************************");
         System.out.println("Buy Price     : ");
-        double buyPrice = in.nextInt(); 
+        double buyPrice = in.nextDouble(); 
         System.out.println("Stop Loss %   : ");
-        double stopLoss = in.nextInt(); 
+        double stopLoss = in.nextDouble(); 
         System.out.println("Profit Book % : ");
-        double profitBook = in.nextInt();
+        double profitBook = in.nextDouble();
         int status = 1;
         int success = db.insertStockTrigger(symbol, buyPrice, stopLoss, profitBook, status);
         if (success == 1)
